requirements.txt
anyio==3.6.2
asgiref==3.6.0
backoff==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opentelemetry-api==1.18.0
opentelemetry-distro==0.39b0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-aws-lambda==0.39b0
opentelemetry-instrumentation-dbapi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-instrumentation-jinja2==0.39b0
opentelemetry-instrumentation-logging==0.39b0
opentelemetry-instrumentation-psycopg2==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-sqlalchemy==0.39b0
opentelemetry-instrumentation-sqlite3==0.39b0
opentelemetry-instrumentation-tortoiseorm==0.39b0
opentelemetry-instrumentation-urllib==0.39b0
opentelemetry-instrumentation-urllib3==0.39b0
opentelemetry-instrumentation-wsgi==0.39b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
orjson==3.8.12
packaging==23.1
pandas==2.0.1
protobuf==4.23.1
psycopg2==2.9.6
pydantic==1.10.7
pydotenv==0.0.7
pymssql==2.2.7
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.12
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
zipp==3.15.0
